requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.4.2
python-dotenv==1.0.0
pymongo==4.5.0
pytest==7.4.3
httpx==0.25.1
websockets==12.0
PyJWT==2.8.0
python-decouple==3.8
email-validator==2.1.0.post1